feat(M008/S05): Decompose index.ts from 2416 to 696 lines
Extract 5 focused modules:
- dashboard.ts (323 lines) — dashboard rendering, welcome screen
- file-handling.ts (243 lines) — paste, drop, file attachments
- message-handler.ts (884 lines) — extension event handler
- keyboard.ts (304 lines) — keyboard, click handlers, ARIA
- ui-updates.ts (362 lines) — header, footer, input, overlay UI

All 140 tests pass. Zero behavioral changes.
