+++ src/webview/index.ts
@@ -5,31 +5,17 @@
 
 import type {
   WebviewToExtensionMessage,
-  ExtensionToWebviewMessage,
-  ProcessStatus,
-  WorkflowState,
-  DashboardData,
-  DashboardSlice,
 } from "../shared/types";
 import "./styles.css";
 
 import {
   state,
   nextId,
-  type ChatEntry,
-  type ToolCallState,
 } from "./state";
 
 import {
-  escapeHtml,
-  formatCost,
-  formatTokens,
-  formatContextUsage,
   formatRelativeTime,
-  formatMarkdownNotes,
-  formatShortDate,
   scrollToBottom,
-  sanitizeUrl,
 } from "./helpers";
 
 import * as slashMenu from "./slash-menu";
@@ -39,6 +25,11 @@ import * as sessionHistory from "./session-history";
 import * as uiDialogs from "./ui-dialogs";
 import * as toasts from "./toasts";
 import * as renderer from "./renderer";
+import * as dashboard from "./dashboard";
+import * as fileHandling from "./file-handling";
+import * as messageHandler from "./message-handler";
+import * as keyboard from "./keyboard";
+import * as uiUpdates from "./ui-updates";
 
 // VS Code API
 declare function acquireVsCodeApi(): {
@@ -91,7 +82,7 @@ function handleToolWatchdogTimeout(toolCallId: string): void {
   renderer.updateToolSegmentElement(toolCallId);
 
   // Show a system message
-  addSystemEntry(
+  messageHandler.addSystemEntry(
     `Tool "${tc.name}" timed out after ${TOOL_WATCHDOG_TIMEOUT_MS / 1000}s. You can abort the current operation or force-restart GSD.`,
     "warning"
   );
@@ -393,212 +384,7 @@ welcomeActions.addEventListener("click", (e: Event) => {
   sendMessage();
 });
 
-// ============================================================
-// Image paste & drop
-// ============================================================
-
-function handleFiles(files: FileList | File[]): void {
-  for (const file of Array.from(files)) {
-    if (file.type.startsWith("image/")) {
-      // Images → inline preview + base64 attachment
-      const reader = new FileReader();
-      reader.onload = () => {
-        const base64 = (reader.result as string).split(",")[1];
-        state.images.push({ type: "image", data: base64, mimeType: file.type });
-        renderImagePreviews();
-      };
-      reader.readAsDataURL(file);
-    } else {
-      // Non-image files (PDFs, docs, etc.) → save to temp, insert path
-      const reader = new FileReader();
-      reader.onload = () => {
-        const base64 = (reader.result as string).split(",")[1];
-        vscode.postMessage({
-          type: "save_temp_file",
-          name: file.name,
-          data: base64,
-          mimeType: file.type,
-        });
-      };
-      reader.readAsDataURL(file);
-    }
-  }
-}
-
-document.addEventListener("paste", (e: ClipboardEvent) => {
-  const items = e.clipboardData?.items;
-  if (!items) return;
-  let handled = false;
-  for (let i = 0; i < items.length; i++) {
-    const item = items[i];
-    if (item.kind === "file") {
-      const file = item.getAsFile();
-      if (file) {
-        handleFiles([file]);
-        handled = true;
-      }
-    }
-  }
-  if (handled) e.preventDefault();
-});
-
-const inputArea = root.querySelector(".gsd-input-area")!;
-inputArea.addEventListener("dragover", (e: Event) => {
-  e.preventDefault();
-  (e as DragEvent).dataTransfer!.dropEffect = "copy";
-  inputArea.classList.add("drag-over");
-});
-inputArea.addEventListener("dragleave", () => inputArea.classList.remove("drag-over"));
-inputArea.addEventListener("drop", (e: Event) => {
-  e.preventDefault();
-  inputArea.classList.remove("drag-over");
-  const dt = (e as DragEvent).dataTransfer;
-  if (!dt) return;
-
-  // Check for file URIs (VS Code explorer drops, OS file manager drops)
-  const uriList = dt.getData("text/uri-list");
-  if (uriList) {
-    const paths = parseDroppedUris(uriList);
-    if (paths.length > 0) {
-      insertDroppedPaths(paths);
-      return;
-    }
-  }
-
-  // Check for plain text paths (some sources use text/plain)
-  const plainText = dt.getData("text/plain");
-  if (plainText && !dt.files.length) {
-    // Heuristic: looks like file path(s) — absolute paths or backslash paths
-    const lines = plainText.split(/\r?\n/).map(l => l.trim()).filter(Boolean);
-    const looksLikePaths = lines.every(l =>
-      /^[A-Z]:\\/.test(l) || l.startsWith("/") || l.startsWith("~")
-    );
-    if (looksLikePaths) {
-      insertDroppedPaths(lines);
-      return;
-    }
-  }
-
-  // Fall through to image handling
-  if (dt.files.length) handleFiles(dt.files);
-});
-
-/** Parse file:// URIs from a text/uri-list drop payload into local paths */
-function parseDroppedUris(uriList: string): string[] {
-  const paths: string[] = [];
-  for (const line of uriList.split(/\r?\n/)) {
-    const trimmed = line.trim();
-    if (!trimmed || trimmed.startsWith("#")) continue;
-    if (trimmed.startsWith("file://")) {
-      try {
-        // file:///C:/foo or file:///home/user/foo
-        const url = new URL(trimmed);
-        let fsPath = decodeURIComponent(url.pathname);
-        // On Windows, pathname is /C:/foo — strip leading slash
-        if (/^\/[A-Za-z]:\//.test(fsPath)) {
-          fsPath = fsPath.slice(1);
-        }
-        paths.push(fsPath);
-      } catch {
-        // Malformed URI — skip
-      }
-    }
-  }
-  return paths;
-}
-
-/** Add file paths as file attachment chips */
-function addFileAttachments(paths: string[], autoSend = false): void {
-  for (const p of paths) {
-    const normalized = p.replace(/\\/g, "/");
-    const parts = normalized.split("/");
-    const name = parts[parts.length - 1] || p;
-    const extMatch = name.match(/\.([^.]+)$/);
-    const extension = extMatch ? extMatch[1].toLowerCase() : "";
-    // Avoid duplicates
-    if (!state.files.some(f => f.path === p)) {
-      state.files.push({ type: "file", path: p, name, extension });
-    }
-  }
-  renderFileChips();
-
-  // Check read access
-  vscode.postMessage({ type: "check_file_access", paths });
-
-  if (autoSend) {
-    sendMessage();
-  }
-}
-
-function getFileIcon(ext: string): string {
-  const icons: Record<string, string> = {
-    pdf: "📄", doc: "📝", docx: "📝", txt: "📝", md: "📝",
-    xls: "📊", xlsx: "📊", csv: "📊",
-    ppt: "📽️", pptx: "📽️",
-    jpg: "🖼️", jpeg: "🖼️", png: "🖼️", gif: "🖼️", svg: "🖼️", webp: "🖼️",
-    mp4: "🎬", mov: "🎬", avi: "🎬", mkv: "🎬",
-    mp3: "🎵", wav: "🎵", flac: "🎵",
-    zip: "📦", tar: "📦", gz: "📦", rar: "📦", "7z": "📦",
-    js: "⚡", ts: "⚡", jsx: "⚡", tsx: "⚡",
-    py: "🐍", rb: "💎", go: "🔷", rs: "🦀",
-    html: "🌐", css: "🎨", scss: "🎨",
-    json: "📋", yaml: "📋", yml: "📋", toml: "📋", xml: "📋",
-    sh: "⚙️", bash: "⚙️", ps1: "⚙️", cmd: "⚙️", bat: "⚙️",
-    sql: "🗃️", db: "🗃️",
-    env: "🔒", key: "🔒", pem: "🔒",
-  };
-  return icons[ext] || "📎";
-}
-
-function renderFileChips(): void {
-  const container = document.getElementById("fileChips")!;
-  if (state.files.length === 0) {
-    container.style.display = "none";
-    container.innerHTML = "";
-    return;
-  }
-  container.style.display = "flex";
-  container.innerHTML = state.files.map((f, i) => `
-    <div class="gsd-file-chip" title="${escapeHtml(f.path)}">
-      <span class="gsd-file-chip-icon">${getFileIcon(f.extension)}</span>
-      <span class="gsd-file-chip-name">${escapeHtml(f.name)}</span>
-      <button class="gsd-file-chip-remove" data-idx="${i}">×</button>
-    </div>
-  `).join("");
-
-  container.querySelectorAll(".gsd-file-chip-remove").forEach((btn) => {
-    btn.addEventListener("click", (e) => {
-      e.stopPropagation();
-      const idx = parseInt((btn as HTMLElement).dataset.idx!);
-      state.files.splice(idx, 1);
-      renderFileChips();
-    });
-  });
-}
-
-function renderImagePreviews(): void {
-  if (state.images.length === 0) {
-    imagePreview.style.display = "none";
-    imagePreview.innerHTML = "";
-    return;
-  }
-  imagePreview.style.display = "flex";
-  imagePreview.innerHTML = state.images.map((img, i) => `
-    <div class="gsd-image-thumb">
-      <img src="data:${img.mimeType};base64,${img.data}" alt="Attached" />
-      <button class="gsd-image-remove" data-idx="${i}">×</button>
-    </div>
-  `).join("");
-
-  imagePreview.querySelectorAll(".gsd-image-remove").forEach((btn) => {
-    btn.addEventListener("click", (e) => {
-      e.stopPropagation();
-      const idx = parseInt((btn as HTMLElement).dataset.idx!);
-      state.images.splice(idx, 1);
-      renderImagePreviews();
-    });
-  });
-}
+// File & image handling is in file-handling.ts
 
 // ============================================================
 // Slash command menu — input listener
@@ -695,7 +481,7 @@ function sendMessage(): void {
 
     promptInput.value = "";
     state.images = [];
-    renderImagePreviews();
+    fileHandling.renderImagePreviews();
     autoResize();
     return;
   }
@@ -726,1604 +512,98 @@ function sendMessage(): void {
   promptInput.value = "";
   state.images = [];
   state.files = [];
-  renderImagePreviews();
-  renderFileChips();
+  fileHandling.renderImagePreviews();
+  fileHandling.renderFileChips();
   autoResize();
 }
 
-// ============================================================
-// Keyboard handling
-// ============================================================
+// Keyboard & click handlers are in keyboard.ts
 
-promptInput.addEventListener("keydown", (e: KeyboardEvent) => {
-  if (sessionHistory.isVisible()) {
-    if (sessionHistory.handleKeyDown(e)) return;
-  }
-  if (slashMenu.isVisible()) {
-    if (e.key === "ArrowDown") {
-      e.preventDefault();
-      slashMenu.navigateDown();
-      return;
-    }
-    if (e.key === "ArrowUp") {
-      e.preventDefault();
-      slashMenu.navigateUp();
-      return;
-    }
-    if (e.key === "Enter" || e.key === "Tab") {
-      e.preventDefault();
-      slashMenu.selectCurrent();
-      return;
-    }
-    if (e.key === "Escape") {
-      e.preventDefault();
-      slashMenu.hide();
-      return;
-    }
-  }
 
-  if (state.useCtrlEnterToSend) {
-    if ((e.ctrlKey || e.metaKey) && e.key === "Enter") {
-      e.preventDefault();
-      sendMessage();
-    }
-  } else {
-    if (e.key === "Enter" && !e.shiftKey) {
-      e.preventDefault();
-      sendMessage();
-    }
-  }
-  if (e.key === "Escape" && state.isStreaming) {
-    vscode.postMessage({ type: "interrupt" });
-  }
-});
 
-// Global keyboard handler for overlay panels
-document.addEventListener("keydown", (e: KeyboardEvent) => {
-  if (sessionHistory.isVisible()) {
-    if (sessionHistory.handleKeyDown(e)) return;
-  }
-  if (e.key === "Escape") {
-    if (thinkingPicker.isVisible()) {
-      thinkingPicker.hide();
-      return;
-    }
-    if (modelPicker.isVisible()) {
-      modelPicker.hide();
-      return;
-    }
-  }
-});
+// UI update functions are in ui-updates.ts
+const { updateAllUI, updateHeaderUI, updateFooterUI, updateInputUI, updateOverlayIndicators, updateWorkflowBadge } = uiUpdates;
 
-headerVersion.addEventListener("click", () => {
-  const existing = document.getElementById("gsd-changelog");
-  if (existing) {
-    existing.classList.add("dismissing");
-    setTimeout(() => existing.remove(), 300);
-  } else {
-    // Show loading spinner while fetching
-    const loader = document.createElement("div");
-    loader.id = "gsd-changelog";
-    loader.className = "gsd-changelog";
-    loader.innerHTML = `
-      <div class="gsd-changelog-header">
-        <span class="gsd-changelog-title">📋 Changelog</span>
-      </div>
-      <div class="gsd-loading-spinner"><div class="gsd-spinner"></div> Loading...</div>
-    `;
-    messagesContainer.appendChild(loader);
-    scrollToBottom(messagesContainer, true);
-    vscode.postMessage({ type: "get_changelog" } as WebviewToExtensionMessage);
-  }
-});
+// Dashboard functions are in dashboard.ts
 
-// Keyboard support for version badge (role="button")
-headerVersion.addEventListener("keydown", (e: KeyboardEvent) => {
-  if (e.key === "Enter" || e.key === " ") {
-    e.preventDefault();
-    headerVersion.click();
-  }
-});
 
-sendBtn.addEventListener("click", () => {
-  if (state.isStreaming) {
-    vscode.postMessage({ type: "interrupt" });
-  } else {
-    sendMessage();
-  }
-});
 
-function handleNewConversation(): void {
-  vscode.postMessage({ type: "new_conversation" });
-  state.entries = [];
-  state.currentTurn = null;
-  renderer.resetStreamingState();
-  state.sessionStats = {};
-  renderer.clearMessages();
-  welcomeScreen.style.display = "flex";
-  sessionHistory.hide();
-  modelPicker.hide();
-  updateAllUI();
-}
-
-newConvoBtn.addEventListener("click", handleNewConversation);
-
-compactBtn.addEventListener("click", () => {
-  if (!state.isStreaming) {
-    vscode.postMessage({ type: "compact_context" });
-    toasts.show("Compacting context…");
-  }
-});
-
-exportBtn.addEventListener("click", () => {
-  vscode.postMessage({ type: "export_html" });
-  toasts.show("Exporting conversation…");
-});
-
-attachBtn.addEventListener("click", () => {
-  vscode.postMessage({ type: "attach_files" });
-});
-
-// Thinking badge click is handled by thinkingPicker.init()
-thinkingBadge.style.cursor = "pointer";
-
-// ============================================================
-// Global click handlers (copy, file links, url links, tool toggles)
-// ============================================================
-
-document.addEventListener("click", (e: Event) => {
-  const target = e.target as HTMLElement;
-
-  const copyBtn = target.closest(".gsd-copy-btn") as HTMLElement | null;
-  if (copyBtn) {
-    const codeBlock = copyBtn.closest(".gsd-code-block");
-    const code = codeBlock?.querySelector("code")?.textContent || "";
-    vscode.postMessage({ type: "copy_text", text: code });
-    copyBtn.textContent = "✓ Copied";
-    setTimeout(() => { copyBtn.textContent = "Copy"; }, 1500);
-    return;
-  }
-
-  const copyResponseBtn = target.closest(".gsd-copy-response-btn") as HTMLElement | null;
-  if (copyResponseBtn) {
-    const text = copyResponseBtn.dataset.copyText || "";
-    vscode.postMessage({ type: "copy_text", text });
-    toasts.show("Response copied");
-    return;
-  }
-
-  if (target.classList.contains("gsd-file-link")) {
-    const path = target.dataset.path;
-    if (path) vscode.postMessage({ type: "open_file", path });
-    return;
-  }
-
-  if (target.tagName === "A" && target.getAttribute("href")) {
-    e.preventDefault();
-    const href = target.getAttribute("href")!;
-    const safeUrl = sanitizeUrl(href);
-    if (safeUrl) {
-      vscode.postMessage({ type: "open_url", url: safeUrl });
-    }
-    return;
-  }
-
-  const toolHeader = target.closest(".gsd-tool-header") as HTMLElement | null;
-  if (toolHeader) {
-    const block = toolHeader.closest(".gsd-tool-block") as HTMLElement | null;
-    if (block) {
-      block.classList.toggle("collapsed");
-      const isExpanded = !block.classList.contains("collapsed");
-      toolHeader.setAttribute("aria-expanded", String(isExpanded));
-    }
-    return;
-  }
-
-  if (target.closest("#restartBtn")) {
-    vscode.postMessage({ type: "launch_gsd" });
-    return;
-  }
-});
-
-// Keyboard activation for role="button" elements (tool headers, group headers)
-messagesContainer.addEventListener("keydown", (e: KeyboardEvent) => {
-  if (e.key !== "Enter" && e.key !== " ") return;
-  const target = e.target as HTMLElement;
-  if (target.getAttribute("role") !== "button") return;
-  e.preventDefault();
-  target.click();
-});
-
-// Sync aria-expanded on <details> toggle for group headers
-messagesContainer.addEventListener("toggle", (e: Event) => {
-  const details = e.target as HTMLDetailsElement;
-  if (!details.classList.contains("gsd-tool-group")) return;
-  const summary = details.querySelector(".gsd-tool-group-header");
-  if (summary) {
-    summary.setAttribute("aria-expanded", String(details.open));
-  }
-}, true);
-
-document.addEventListener("copy", (_e: ClipboardEvent) => {
-  const selection = window.getSelection()?.toString();
-  if (selection) vscode.postMessage({ type: "copy_text", text: selection });
-});
-
-
-
-// ============================================================
-// UI Updates
-// ============================================================
-
-function updateAllUI(): void {
-  updateHeaderUI();
-  updateFooterUI();
-  updateInputUI();
-  updateOverlayIndicators();
-  updateWelcomeScreen();
-}
-
-function updateHeaderUI(): void {
-  if (state.model) {
-    modelBadge.textContent = state.model.name || state.model.id;
-    modelBadge.title = `${state.model.provider} / ${state.model.id}`;
-    modelBadge.style.display = "inline-flex";
-  } else {
-    modelBadge.style.display = "none";
-  }
-
-  // Thinking badge — model-aware
-  const modelSupportsReasoning = state.model
-    ? state.availableModels.some(
-        (m) => m.id === state.model!.id && m.provider === state.model!.provider && m.reasoning
-      )
-    : false;
-
-  if (state.model && !modelSupportsReasoning && state.modelsLoaded) {
-    // Non-reasoning model — show disabled badge
-    thinkingBadge.textContent = "🧠 N/A";
-    thinkingBadge.title = "Current model does not support extended thinking";
-    thinkingBadge.style.display = "inline-flex";
-    thinkingBadge.classList.add("disabled");
-  } else {
-    const thinkingLabel = state.thinkingLevel && state.thinkingLevel !== "off" ? state.thinkingLevel : "off";
-    thinkingBadge.textContent = `🧠 ${thinkingLabel}`;
-    thinkingBadge.title = "Click to select thinking level";
-    thinkingBadge.style.display = "inline-flex";
-    thinkingBadge.classList.remove("disabled");
-  }
-
-  const stats = state.sessionStats;
-  const hasCost = stats.cost != null && stats.cost > 0;
-  if (hasCost) {
-    costBadge.textContent = formatCost(stats.cost);
-    costBadge.style.display = "inline-flex";
-  } else {
-    costBadge.style.display = "none";
-  }
-
-  const ctx = formatContextUsage(stats, state.model);
-  if (ctx) {
-    contextBadge.textContent = `◐ ${ctx}`;
-    contextBadge.style.display = "inline-flex";
-    const pct = stats.contextPercent ?? 0;
-    contextBadge.classList.remove("warn", "crit");
-    if (pct > 90) contextBadge.classList.add("crit");
-    else if (pct > 70) contextBadge.classList.add("warn");
-  } else {
-    contextBadge.style.display = "none";
-  }
-
-  // Show separator between model/thinking and cost/context groups
-  const hasLeftBadges = modelBadge.style.display !== "none" || thinkingBadge.style.display !== "none";
-  const hasRightBadges = costBadge.style.display !== "none" || contextBadge.style.display !== "none";
-  headerSep1.style.display = hasLeftBadges && hasRightBadges ? "block" : "none";
-
-  // Context usage bar
-  updateContextBar();
-}
-
-function updateContextBar(): void {
-  const pct = state.sessionStats.contextPercent ?? 0;
-  if (pct <= 0) {
-    contextBarContainer.style.display = "none";
-    return;
-  }
-
-  contextBarContainer.style.display = "block";
-  contextBar.style.width = `${Math.min(pct, 100)}%`;
-
-  contextBar.classList.remove("ok", "warn", "crit");
-  if (pct > 90) {
-    contextBar.classList.add("crit");
-  } else if (pct > 70) {
-    contextBar.classList.add("warn");
-  } else {
-    contextBar.classList.add("ok");
-  }
-}
-
-function updateFooterUI(): void {
-  footerCwd.textContent = state.cwd || "";
-  footerCwd.title = state.cwd;
-
-  const stats = state.sessionStats;
-  const tokens = stats.tokens;
-  const parts: string[] = [];
-
-  if (tokens) {
-    const tokenParts: string[] = [];
-    if (tokens.input) tokenParts.push(`in:${formatTokens(tokens.input)}`);
-    if (tokens.output) tokenParts.push(`out:${formatTokens(tokens.output)}`);
-    if (tokens.cacheRead) tokenParts.push(`cache:${formatTokens(tokens.cacheRead)}`);
-    if (tokenParts.length > 0) parts.push(tokenParts.join(" "));
-  }
-
-  if (stats.cost != null && stats.cost > 0) {
-    parts.push(formatCost(stats.cost));
-  }
-
-  const ctx = formatContextUsage(stats, state.model);
-  if (ctx) parts.push(ctx);
-
-  footerStats.textContent = parts.join(" · ");
-
-  let right = "";
-  if (state.model) {
-    right = state.model.id || state.model.name || "";
-    if (state.thinkingLevel && state.thinkingLevel !== "off") {
-      right += ` • ${state.thinkingLevel}`;
-    } else {
-      right += " • thinking off";
-    }
-  }
-  footerRight.textContent = right;
-}
-
-function updateInputUI(): void {
-  if (state.isStreaming) {
-    sendIcon.textContent = "■";
-    sendBtn.classList.add("gsd-stop-btn");
-    sendBtn.title = "Stop (Esc)";
-    promptInput.placeholder = "Interrupt or steer GSD...";
-  } else {
-    sendIcon.textContent = "↑";
-    sendBtn.classList.remove("gsd-stop-btn");
-    sendBtn.title = "Send";
-    promptInput.placeholder = "Message GSD...";
-  }
-
-  const sendKey = state.useCtrlEnterToSend ? "Ctrl+Enter" : "Enter";
-  if (state.isStreaming) {
-    inputHint.textContent = `${sendKey} to steer • Esc to stop`;
-  } else {
-    inputHint.textContent = `${sendKey} to send • !cmd for bash • / for commands`;
-  }
-}
-
-function updateOverlayIndicators(): void {
-  const parts: string[] = [];
-
-  if (state.isCompacting) {
-    parts.push(`<div class="gsd-overlay-indicator compacting">
-      <span class="gsd-overlay-spinner"></span>Compacting context…
-    </div>`);
-  }
-
-  if (state.isRetrying && state.retryInfo) {
-    parts.push(`<div class="gsd-overlay-indicator retrying">
-      <span class="gsd-overlay-spinner"></span>Retrying (${state.retryInfo.attempt}/${state.retryInfo.maxAttempts})…
-      <span class="gsd-overlay-detail">${escapeHtml(state.retryInfo.errorMessage)}</span>
-    </div>`);
-  }
-
-  if (state.processStatus === "starting") {
-    parts.push(`<div class="gsd-overlay-indicator starting">
-      <span class="gsd-overlay-spinner"></span>Starting GSD…
-    </div>`);
-  }
-
-  if (state.processStatus === "restarting") {
-    parts.push(`<div class="gsd-overlay-indicator restarting">
-      <span class="gsd-overlay-spinner"></span>Restarting GSD…
-    </div>`);
-  }
-
-  if (state.processStatus === "crashed") {
-    parts.push(`<div class="gsd-overlay-indicator crashed">
-      ⚠ GSD process not running
-      <button id="restartBtn" class="gsd-overlay-btn">Restart</button>
-    </div>`);
-  }
-
-  if (state.processHealth === "unresponsive") {
-    parts.push(`<div class="gsd-overlay-indicator unresponsive">
-      ⚠ GSD is unresponsive
-      <button id="forceRestartBtn" class="gsd-overlay-btn danger">Force Restart</button>
-      <button id="forceKillBtn" class="gsd-overlay-btn">Kill Process</button>
-    </div>`);
-  }
-
-  overlayIndicators.innerHTML = parts.join("");
-  overlayIndicators.style.display = parts.length > 0 ? "flex" : "none";
-
-  // Wire up force buttons (if rendered)
-  const forceRestartBtn = document.getElementById("forceRestartBtn");
-  if (forceRestartBtn) {
-    forceRestartBtn.addEventListener("click", () => {
-      vscode.postMessage({ type: "force_restart" });
-      state.processHealth = "responsive";
-      state.processStatus = "restarting";
-      updateOverlayIndicators();
-    });
-  }
-  const forceKillBtn = document.getElementById("forceKillBtn");
-  if (forceKillBtn) {
-    forceKillBtn.addEventListener("click", () => {
-      vscode.postMessage({ type: "force_kill" });
-      state.processHealth = "responsive";
-      updateOverlayIndicators();
-    });
-  }
-}
-
-function updateWorkflowBadge(wf: WorkflowState | null): void {
-  const badge = document.getElementById("workflowBadge");
-  if (!badge) return;
-
-  if (!wf) {
-    badge.textContent = "Self-directed";
-    badge.className = "gsd-workflow-badge";
-    badge.style.display = "inline-flex";
-    return;
-  }
-
-  const parts: string[] = [];
-
-  // Build breadcrumb: M004 › S02 › T03
-  if (wf.milestone) parts.push(wf.milestone.id);
-  if (wf.slice) parts.push(wf.slice.id);
-  if (wf.task) parts.push(wf.task.id);
-
-  // Phase label
-  const phaseLabels: Record<string, string> = {
-    "pre-planning": "Pre-planning",
-    "discussing": "Discussing",
-    "researching": "Researching",
-    "planning": "Planning",
-    "executing": "Executing",
-    "verifying": "Verifying",
-    "summarizing": "Summarizing",
-    "advancing": "Advancing",
-    "completing-milestone": "Completing",
-    "replanning-slice": "Replanning",
-    "complete": "Complete",
-    "paused": "Paused",
-    "blocked": "Blocked",
-    "unknown": "",
-  };
-  const phaseText = phaseLabels[wf.phase] || wf.phase;
-
-  // Build display text
-  let text: string;
-  if (parts.length > 0) {
-    text = parts.join(" › ");
-    if (phaseText && phaseText !== "Complete") {
-      text += ` · ${phaseText}`;
-    } else if (phaseText === "Complete") {
-      text += " ✓";
-    }
-  } else if (phaseText) {
-    text = phaseText;
-  } else {
-    text = "Self-directed";
-  }
-
-  // Auto-mode prefix
-  if (wf.autoMode === "auto") {
-    text = `⚡ ${text}`;
-  } else if (wf.autoMode === "next") {
-    text = `▸ ${text}`;
-  } else if (wf.autoMode === "paused") {
-    text = `⏸ ${text}`;
-  }
-
-  badge.textContent = text;
-
-  // Phase-based styling
-  let extraClass = "";
-  if (wf.phase === "blocked") extraClass = " blocked";
-  else if (wf.phase === "paused") extraClass = " paused";
-  else if (wf.phase === "complete") extraClass = " complete";
-  else if (wf.autoMode) extraClass = " auto";
-
-  badge.className = `gsd-workflow-badge${extraClass}`;
-  badge.style.display = "inline-flex";
-}
-
-// ============================================================
-// Dashboard Panel
-// ============================================================
-
-function renderDashboard(data: DashboardData | null): void {
-  welcomeScreen.style.display = "none";
-
-  // Remove any existing dashboard
-  const existing = document.querySelector(".gsd-dashboard");
-  if (existing) existing.remove();
-
-  const el = document.createElement("div");
-  el.className = "gsd-dashboard";
-
-  if (!data || (!data.hasProject && !data.hasMilestone)) {
-    el.innerHTML = `
-      <div class="gsd-dashboard-empty">
-        <div class="gsd-dashboard-empty-icon">📊</div>
-        <div class="gsd-dashboard-empty-text">No active GSD project</div>
-        <div class="gsd-dashboard-empty-hint">Run <code>/gsd</code> to start a project in this workspace</div>
-      </div>
-    `;
-    messagesContainer.appendChild(el);
-    scrollToBottom(messagesContainer, true);
-    return;
-  }
-
-  const phaseLabels: Record<string, string> = {
-    "pre-planning": "Pre-planning",
-    "discussing": "Discussing",
-    "researching": "Researching",
-    "planning": "Planning",
-    "executing": "Executing",
-    "verifying": "Verifying",
-    "summarizing": "Summarizing",
-    "advancing": "Advancing",
-    "completing-milestone": "Completing",
-    "replanning-slice": "Replanning",
-    "complete": "Complete",
-    "paused": "Paused",
-    "blocked": "Blocked",
-    "unknown": "",
-  };
-
-  const phaseText = phaseLabels[data.phase] || data.phase;
-  const phaseClass = data.phase === "complete" ? "complete"
-    : data.phase === "blocked" ? "blocked"
-    : data.phase === "executing" ? "executing" : "";
-
-  // Build current action breadcrumb
-  const breadcrumb: string[] = [];
-  if (data.milestone) breadcrumb.push(data.milestone.id);
-  if (data.slice) breadcrumb.push(data.slice.id);
-  if (data.task) breadcrumb.push(data.task.id);
-
-  // Progress bar helper
-  function progressBar(done: number, total: number, label: string): string {
-    if (total === 0) return "";
-    const pct = Math.round((done / total) * 100);
-    const fillPct = (done / total) * 100;
-    return `
-      <div class="gsd-dash-progress-row">
-        <span class="gsd-dash-progress-label">${escapeHtml(label)}</span>
-        <div class="gsd-dash-progress-track">
-          <div class="gsd-dash-progress-fill" style="width: ${fillPct}%"></div>
-        </div>
-        <span class="gsd-dash-progress-pct">${pct}%</span>
-        <span class="gsd-dash-progress-ratio">${done}/${total}</span>
-      </div>
-    `;
-  }
-
-  // Slice list with nested tasks
-  function sliceList(slices: DashboardSlice[]): string {
-    if (slices.length === 0) return "";
-    return `<div class="gsd-dash-slices">` + slices.map(s => {
-      const icon = s.done ? "✓" : s.active ? "▸" : "○";
-      const cls = s.done ? "done" : s.active ? "active" : "pending";
-      const riskCls = `risk-${s.risk}`;
-
-      let tasksHtml = "";
-      if (s.active && s.tasks.length > 0) {
-        tasksHtml = `<div class="gsd-dash-tasks">` +
-          s.tasks.map(t => {
-            const tIcon = t.done ? "✓" : t.active ? "▸" : "·";
-            const tCls = t.done ? "done" : t.active ? "active" : "pending";
-            return `<div class="gsd-dash-task ${tCls}"><span class="gsd-dash-icon">${tIcon}</span> ${escapeHtml(t.id)}: ${escapeHtml(t.title)}</div>`;
-          }).join("") +
-          `</div>`;
-      }
-
-      return `
-        <div class="gsd-dash-slice ${cls}">
-          <div class="gsd-dash-slice-row">
-            <span class="gsd-dash-icon">${icon}</span>
-            <span class="gsd-dash-slice-title">${escapeHtml(s.id)}: ${escapeHtml(s.title)}</span>
-            <span class="gsd-dash-risk ${riskCls}">${escapeHtml(s.risk)}</span>
-          </div>
-          ${tasksHtml}
-        </div>
-      `;
-    }).join("") + `</div>`;
-  }
-
-  // Milestone registry
-  function milestoneList(entries: typeof data.milestoneRegistry): string {
-    if (entries.length === 0) return "";
-    return `
-      <div class="gsd-dash-section">
-        <div class="gsd-dash-section-title">Milestones</div>
-        <div class="gsd-dash-milestones">
-          ${entries.map(m => {
-            const icon = m.done ? "✓" : m.active ? "▸" : "○";
-            const cls = m.done ? "done" : m.active ? "active" : "pending";
-            return `<div class="gsd-dash-milestone ${cls}"><span class="gsd-dash-icon">${icon}</span> ${escapeHtml(m.id)}: ${escapeHtml(m.title)}</div>`;
-          }).join("")}
-        </div>
-      </div>
-    `;
-  }
-
-  // Cost & usage section
-  function costSection(stats: typeof data.stats): string {
-    if (!stats) return "";
-    const parts: string[] = [];
-
-    if (stats.cost != null) {
-      parts.push(`<span class="gsd-dash-cost-value">$${stats.cost.toFixed(4)}</span> total`);
-    }
-    if (stats.tokens?.total) {
-      parts.push(`${formatTokenCount(stats.tokens.total)} tokens`);
-    }
-    if (stats.toolCalls) {
-      parts.push(`${stats.toolCalls} tools`);
-    }
-    if (stats.userMessages) {
-      parts.push(`${stats.userMessages} turns`);
-    }
-
-    if (parts.length === 0) return "";
-
-    let tokenDetail = "";
-    if (stats.tokens) {
-      const t = stats.tokens;
-      tokenDetail = `<div class="gsd-dash-cost-detail">in: ${formatTokenCount(t.input)}  out: ${formatTokenCount(t.output)}  cache-r: ${formatTokenCount(t.cacheRead)}  cache-w: ${formatTokenCount(t.cacheWrite)}</div>`;
-    }
-
-    return `
-      <div class="gsd-dash-section">
-        <div class="gsd-dash-section-title">Cost & Usage</div>
-        <div class="gsd-dash-cost-summary">${parts.join("  ·  ")}</div>
-        ${tokenDetail}
-      </div>
-    `;
-  }
-
-  // Blockers section
-  function blockersSection(blockers: string[]): string {
-    if (blockers.length === 0) return "";
-    return `
-      <div class="gsd-dash-section">
-        <div class="gsd-dash-section-title gsd-dash-blockers-title">⚠ Blockers</div>
-        ${blockers.map(b => `<div class="gsd-dash-blocker">${escapeHtml(b)}</div>`).join("")}
-      </div>
-    `;
-  }
-
-  // Next action
-  function nextActionSection(next: string | null): string {
-    if (!next) return "";
-    return `
-      <div class="gsd-dash-section">
-        <div class="gsd-dash-section-title">Next</div>
-        <div class="gsd-dash-next">${escapeHtml(next)}</div>
-      </div>
-    `;
-  }
-
-  // Build the full dashboard
-  const hasActiveWork = data.hasMilestone && data.milestone;
-
-  el.innerHTML = `
-    <div class="gsd-dashboard-header">
-      <div class="gsd-dashboard-title">📊 GSD Dashboard</div>
-      <span class="gsd-dashboard-phase ${phaseClass}">${escapeHtml(phaseText)}</span>
-    </div>
-
-    ${(data.task || data.slice) ? `
-    <div class="gsd-dash-current">
-      <span class="gsd-dash-current-label">Now:</span>
-      <span class="gsd-dash-current-action">${escapeHtml(phaseText)} ${escapeHtml(breadcrumb.join("/"))}</span>
-    </div>
-    ` : ""}
-
-    ${hasActiveWork ? `
-      <div class="gsd-dash-milestone-title">${escapeHtml(data.milestone!.id)}: ${escapeHtml(data.milestone!.title)}</div>
-
-      <div class="gsd-dash-progress">
-        ${progressBar(data.progress.tasks.done, data.progress.tasks.total, "Tasks")}
-        ${progressBar(data.progress.slices.done, data.progress.slices.total, "Slices")}
-        ${progressBar(data.progress.milestones.done, data.progress.milestones.total, "Milestones")}
-      </div>
-
-      ${sliceList(data.slices)}
-    ` : `
-      <div class="gsd-dash-progress">
-        ${progressBar(data.progress.milestones.done, data.progress.milestones.total, "Milestones")}
-      </div>
-    `}
-
-    ${milestoneList(data.milestoneRegistry)}
-    ${blockersSection(data.blockers)}
-    ${costSection(data.stats)}
-    ${nextActionSection(data.nextAction)}
-
-    <div class="gsd-dashboard-footer">
-      <span class="gsd-dash-hint">${hasActiveWork ? `Run <code>/gsd auto</code> to start executing` : `Run <code>/gsd</code> to start a new milestone`}</span>
-    </div>
-  `;
-
-  messagesContainer.appendChild(el);
-  scrollToBottom(messagesContainer, true);
-}
-
-function formatTokenCount(n: number): string {
-  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(2)}M`;
-  if (n >= 1_000) return `${(n / 1_000).toFixed(1)}k`;
-  return String(n);
-}
-
-function updateWelcomeScreen(): void {
-  // Only hide welcome for real conversation entries (user messages, assistant turns)
-  // System info/warning entries alone shouldn't dismiss it
-  const hasConversation = state.currentTurn || state.entries.some(
-    e => e.type === "user" || e.type === "assistant"
-  );
-  if (hasConversation) {
-    welcomeScreen.style.display = "none";
-    return;
-  }
-  welcomeScreen.style.display = "flex";
-
-  welcomeVersion.textContent = state.version ? `v${state.version}` : "";
-
-  switch (state.processStatus) {
-    case "starting":
-      welcomeProcess.textContent = "Starting GSD…";
-      break;
-    case "running":
-      welcomeProcess.textContent = "Type a message to start";
-      break;
-    case "crashed":
-      welcomeProcess.textContent = "GSD failed to start — check the error below";
-      break;
-    case "restarting":
-      welcomeProcess.textContent = "Restarting…";
-      break;
-    default:
-      welcomeProcess.textContent = "Initializing…";
-  }
-
-  if (state.model && state.processStatus === "running") {
-    let modelStr = state.model.id || state.model.name || "";
-    if (state.model.provider) modelStr = `${state.model.provider}/${modelStr}`;
-    if (state.thinkingLevel && state.thinkingLevel !== "off") {
-      modelStr += ` • 🧠 ${state.thinkingLevel}`;
-    }
-    welcomeModel.textContent = modelStr;
-    welcomeModel.style.display = "block";
-  } else {
-    welcomeModel.style.display = "none";
-  }
-
-  if (state.processStatus === "running") {
-    const sendKey = state.useCtrlEnterToSend ? "Ctrl+Enter" : "Enter";
-    welcomeHints.innerHTML = [
-      `<span>${sendKey} to send</span>`,
-      `<span>Shift+Enter for newline</span>`,
-      `<span>Esc to interrupt</span>`,
-      `<span>/ for commands</span>`,
-    ].join('<span class="gsd-hint-sep">•</span>');
-    welcomeHints.style.display = "flex";
-  } else {
-    welcomeHints.style.display = "none";
-  }
-}
-
-
-
-// ============================================================
-// Handle messages FROM extension
-// ============================================================
-
-window.addEventListener("message", (event) => {
-  const raw = event.data as Record<string, unknown>;
-  if (!raw || !raw.type) return;
-  const msg = raw as ExtensionToWebviewMessage;
-
-  try {
-
-  switch (msg.type) {
-    case "config": {
-      const data = msg;
-      state.useCtrlEnterToSend = data.useCtrlEnterToSend ?? false;
-      if (data.cwd) state.cwd = data.cwd;
-      if (data.version) state.version = data.version;
-      if (data.extensionVersion) {
-        state.extensionVersion = data.extensionVersion;
-        const headerVer = document.getElementById("headerVersion");
-        if (headerVer) headerVer.textContent = `v${data.extensionVersion}`;
-      }
-      updateAllUI();
-      break;
-    }
-
-    case "state": {
-      const data = msg.data;
-      if (data) {
-        state.model = data.model || null;
-        state.thinkingLevel = data.thinkingLevel || "off";
-        state.isStreaming = data.isStreaming || false;
-        state.isCompacting = data.isCompacting || false;
-        if (data.cwd) state.cwd = data.cwd;
-        if (data.autoCompactionEnabled != null) {
-          state.sessionStats.autoCompactionEnabled = data.autoCompactionEnabled;
-        }
-        if (data.model?.contextWindow) {
-          state.sessionStats.contextWindow = data.model.contextWindow;
-        }
-        if (data.sessionId) {
-          sessionHistory.setCurrentSessionId(data.sessionId);
-        }
-        if (state.processStatus !== "crashed") state.processStatus = "running";
-        // Eagerly fetch available models if not loaded yet (debounce)
-        if (!state.modelsLoaded && !state.modelsRequested) {
-          state.modelsRequested = true;
-          vscode.postMessage({ type: "get_available_models" });
-        }
-        updateAllUI();
-      }
-      break;
-    }
-
-    case "session_stats": {
-      const data = msg.data;
-      if (data) {
-        state.sessionStats = {
-          ...state.sessionStats,
-          ...data,
-        };
-        updateHeaderUI();
-        updateFooterUI();
-      }
-      break;
-    }
-
-    case "process_status": {
-      const data = msg;
-      const prevStatus = state.processStatus;
-      state.processStatus = data.status as ProcessStatus;
-
-      // When the process becomes "running" (fresh start or after crash/restart),
-      // reset command cache so the slash menu re-fetches from the new process.
-      if (data.status === "running" && prevStatus !== "running") {
-        // Reset streaming state — if we're freshly running, we can't be streaming
-        state.isStreaming = false;
-        state.isCompacting = false;
-        state.commandsLoaded = false;
-        state.commands = [];
-        // Eagerly fetch commands so they're ready when the user types /
-        vscode.postMessage({ type: "get_commands" });
-      }
-
-      updateOverlayIndicators();
-      updateWelcomeScreen();
-      break;
-    }
-
-    case "workflow_state": {
-      updateWorkflowBadge(msg.state);
-      break;
-    }
-
-    case "dashboard_data": {
-      renderDashboard(msg.data);
-      break;
-    }
-
-    case "whats_new": {
-      showWhatsNew(msg.version, msg.notes);
-      break;
-    }
-
-    case "changelog": {
-      showChangelog(msg.entries);
-      break;
-    }
-
-    case "agent_start": {
-      // Expire any pending UI dialogs from a previous turn — the backend's
-      // abort signal has already auto-resolved them, so user interaction
-      // would go nowhere.
-      if (uiDialogs.hasPending()) {
-        uiDialogs.expireAllPending("New turn started");
-      }
-      state.isStreaming = true;
-      state.currentTurn = {
-        id: nextId(),
-        segments: [],
-        toolCalls: new Map(),
-        isComplete: false,
-        timestamp: Date.now(),
-      };
-      renderer.resetStreamingState();
-      updateInputUI();
-      renderer.ensureCurrentTurnElement();
-      announceToScreenReader("Assistant is responding...");
-      break;
-    }
-
-    case "agent_end": {
-      state.isStreaming = false;
-      announceToScreenReader("Response complete.");
-      state.processHealth = "responsive";
-      // Clear all tool watchdog timers
-      for (const [, timer] of toolWatchdogTimers) {
-        clearTimeout(timer);
-      }
-      toolWatchdogTimers.clear();
-      // Expire any pending UI dialogs — the backend's abort signal fires
-      // on agent_end, auto-resolving all pending dialogs to defaults.
-      // Mark them so the user sees they're no longer interactive.
-      if (uiDialogs.hasPending()) {
-        uiDialogs.expireAllPending("Agent finished");
-      }
-      renderer.finalizeCurrentTurn();
-      updateInputUI();
-      updateOverlayIndicators();
-      vscode.postMessage({ type: "get_session_stats" });
-      break;
-    }
-
-    case "turn_start": {
-      if (!state.currentTurn) {
-        state.currentTurn = {
-          id: nextId(),
-          segments: [],
-          toolCalls: new Map(),
-          isComplete: false,
-          timestamp: Date.now(),
-        };
-        renderer.resetStreamingState();
-      }
-      break;
-    }
-
-    case "turn_end": {
-      break;
-    }
-
-    case "message_start": {
-      break;
-    }
-
-    case "message_update": {
-      if (!state.currentTurn) break;
-      const data = msg;
-      const delta = data.assistantMessageEvent || data.delta;
-
-      if (delta) {
-        if (delta.type === "text_delta" && delta.delta) {
-          renderer.appendToTextSegment("text", delta.delta);
-        } else if (delta.type === "thinking_delta" && delta.delta) {
-          renderer.appendToTextSegment("thinking", delta.delta);
-        }
-      }
-      break;
-    }
-
-    case "message_end": {
-      const endData = msg;
-      const endMsg = endData.message;
-      if (endMsg?.role === "assistant" && endMsg.usage) {
-        const u = endMsg.usage;
-        if (!state.sessionStats.tokens) {
-          state.sessionStats.tokens = { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 };
-        }
-        const t = state.sessionStats.tokens;
-        t.input += u.input || 0;
-        t.output += u.output || 0;
-        t.cacheRead += u.cacheRead || 0;
-        t.cacheWrite += u.cacheWrite || 0;
-        t.total = t.input + t.output + t.cacheRead + t.cacheWrite;
-        if (u.cost?.total) {
-          state.sessionStats.cost = (state.sessionStats.cost || 0) + u.cost.total;
-        }
-        const contextTokens = (u.input || 0) + (u.cacheRead || 0);
-        const contextWindow = state.model?.contextWindow || state.sessionStats.contextWindow || 0;
-        if (contextWindow > 0 && contextTokens > 0) {
-          state.sessionStats.contextTokens = contextTokens;
-          state.sessionStats.contextWindow = contextWindow;
-          state.sessionStats.contextPercent = (contextTokens / contextWindow) * 100;
-        }
-        updateHeaderUI();
-        updateFooterUI();
-      }
-      break;
-    }
-
-    case "tool_execution_start": {
-      if (!state.currentTurn) break;
-      const data = msg;
-      const tc: ToolCallState = {
-        id: data.toolCallId,
-        name: data.toolName,
-        args: data.args || {},
-        resultText: "",
-        isError: false,
-        isRunning: true,
-        startTime: Date.now(),
-      };
-      state.currentTurn.toolCalls.set(data.toolCallId, tc);
-      const segIdx = state.currentTurn.segments.length;
-      state.currentTurn.segments.push({ type: "tool", toolCallId: data.toolCallId });
-      renderer.appendToolSegmentElement(tc, segIdx);
-      // Start watchdog timer for this tool
-      startToolWatchdog(data.toolCallId);
-      scrollToBottom(messagesContainer);
-      break;
-    }
-
-    case "tool_execution_update": {
-      if (!state.currentTurn) break;
-      const data = msg;
-      const tc = state.currentTurn.toolCalls.get(data.toolCallId);
-      if (tc && data.partialResult) {
-        const text = data.partialResult.content
-          ?.map((c: any) => c.text || "")
-          .filter(Boolean)
-          .join("\n");
-        if (text) tc.resultText = text;
-        renderer.updateToolSegmentElement(data.toolCallId);
-        scrollToBottom(messagesContainer);
-      }
-      break;
-    }
-
-    case "tool_execution_end": {
-      if (!state.currentTurn) break;
-      const data = msg;
-      // Clear watchdog — tool completed normally
-      clearToolWatchdog(data.toolCallId);
-      const tc = state.currentTurn.toolCalls.get(data.toolCallId);
-      if (tc) {
-        tc.isRunning = false;
-        tc.isError = data.isError;
-        tc.endTime = Date.now();
-        if (data.durationMs) {
-          tc.endTime = tc.startTime + data.durationMs;
-        }
-        if (data.result) {
-          const text = data.result.content
-            ?.map((c: any) => c.text || "")
-            .filter(Boolean)
-            .join("\n");
-          if (text) tc.resultText = text;
-        }
-      }
-      renderer.updateToolSegmentElement(data.toolCallId);
-      scrollToBottom(messagesContainer);
-      break;
-    }
-
-    case "auto_compaction_start": {
-      state.isCompacting = true;
-      updateOverlayIndicators();
-      break;
-    }
-
-    case "auto_compaction_end": {
-      state.isCompacting = false;
-      updateOverlayIndicators();
-      break;
-    }
-
-    case "auto_retry_start": {
-      const data = msg;
-      state.isRetrying = true;
-      state.retryInfo = {
-        attempt: data.attempt,
-        maxAttempts: data.maxAttempts,
-        errorMessage: data.errorMessage || "",
-      };
-      updateOverlayIndicators();
-      break;
-    }
-
-    case "auto_retry_end": {
-      const data = msg;
-      state.isRetrying = false;
-      state.retryInfo = undefined;
-      updateOverlayIndicators();
-      if (!data.success && data.finalError) {
-        addSystemEntry(data.finalError, "error");
-      }
-      break;
-    }
-
-    case "extension_error": {
-      const data = msg;
-      const extError = (data as any).error as string || "unknown error";
-      addSystemEntry(`Command error: ${extError}`, "error");
-      break;
-    }
-
-    case "extension_ui_request": {
-      const data = msg;
-      if (data.method === "notify" && data.message) {
-        const notifyType = (data as any).notifyType as string || "info";
-        const kind = notifyType === "error" ? "error" : notifyType === "warning" ? "warning" : "info";
-        addSystemEntry(data.message as string, kind);
-      } else if (data.method === "setStatus" && data.statusText) {
-        // Status text — could update footer
-      } else if (data.method === "set_editor_text" && data.text) {
-        promptInput.value = data.text;
-        autoResize();
-      } else if (data.method === "select" || data.method === "confirm" || data.method === "input") {
-        uiDialogs.handleRequest(data);
-      }
-      break;
-    }
-
-    case "commands": {
-      const data = msg;
-      state.commands = data.commands || [];
-      state.commandsLoaded = true;
-      if (slashMenu.isVisible()) {
-        const filter = promptInput.value.slice(1).trim();
-        slashMenu.show(filter);
-      }
-      break;
-    }
-
-    case "available_models": {
-      const data = msg;
-      state.availableModels = (data.models || []).map((m: any) => ({
-        id: m.id,
-        name: m.name || m.id,
-        provider: m.provider,
-        reasoning: m.reasoning || false,
-        contextWindow: m.contextWindow,
-      }));
-      state.modelsLoaded = true;
-      state.modelsRequested = false;
-      if (modelPicker.isVisible()) {
-        modelPicker.render();
-      }
-      break;
-    }
-
-    case "thinking_level_changed": {
-      const data = msg;
-      state.thinkingLevel = data.level || "off";
-      updateHeaderUI();
-      updateFooterUI();
-      thinkingPicker.refresh();
-      toasts.show(`Thinking: ${state.thinkingLevel}`);
-      break;
-    }
-
-    case "bash_result": {
-      const data = msg;
-      const result = data.result;
-      if (result) {
-        const output = result.stdout || result.stderr || result.output || JSON.stringify(result);
-        const isError = result.exitCode !== 0 || result.error;
-        addSystemEntry(typeof output === "string" ? output : JSON.stringify(output, null, 2), isError ? "error" : "info");
-      }
-      break;
-    }
-
-    case "error": {
-      const data = msg;
-      addSystemEntry(data.message, "error");
-      break;
-    }
-
-    case "process_exit": {
-      const data = msg;
-      state.isStreaming = false;
-      state.isCompacting = false;
-      state.isRetrying = false;
-      state.processHealth = "responsive";
-      state.currentTurn = null;
-      // Clear all tool watchdog timers
-      for (const [, timer] of toolWatchdogTimers) {
-        clearTimeout(timer);
-      }
-      toolWatchdogTimers.clear();
-      // Expire any pending dialogs — the process is gone
-      if (uiDialogs.hasPending()) {
-        uiDialogs.expireAllPending("Process exited");
-      }
-      // Reset command cache — the process that provided them is dead
-      state.commandsLoaded = false;
-      state.commands = [];
-      renderer.resetStreamingState();
-      updateInputUI();
-      updateOverlayIndicators();
-
-      // Build an informative error message including stderr detail
-      const detail = (data as any).detail as string | undefined;
-      let message: string;
-      if (detail) {
-        message = detail;
-      } else if (data.code === 0) {
-        message = "GSD process exited.";
-      } else {
-        message = `GSD process exited (code: ${data.code}).`;
-      }
-      addSystemEntry(message, data.code === 0 ? "info" : "error");
-      break;
-    }
-
-    case "process_health": {
-      const data = msg;
-      state.processHealth = data.status;
-      if (data.status === "unresponsive") {
-        updateOverlayIndicators();
-      } else if (data.status === "recovered") {
-        updateOverlayIndicators();
-        addSystemEntry("GSD process recovered", "info");
-      }
-      break;
-    }
-
-    case "session_list": {
-      const data = msg;
-      sessionHistory.updateSessions(data.sessions || []);
-      break;
-    }
-
-    case "session_list_error": {
-      const data = msg;
-      sessionHistory.showError(data.message);
-      break;
-    }
-
-    case "file_access_result": {
-      const data = msg;
-      const denied = data.results.filter((r: { path: string; readable: boolean }) => !r.readable);
-      if (denied.length > 0) {
-        const names = denied.map((r: { path: string }) => {
-          const parts = r.path.replace(/\\/g, "/").split("/");
-          return parts[parts.length - 1] || r.path;
-        });
-        toasts.show(`⚠ No read access: ${names.join(", ")}`, 4000);
-      }
-      break;
-    }
-
-    case "temp_file_saved": {
-      const data = msg;
-      addFileAttachments([data.path], true);
-      break;
-    }
-
-    case "files_attached": {
-      const data = msg;
-      if (data.paths.length > 0) {
-        addFileAttachments(data.paths, true);
-      }
-      break;
-    }
-
-    case "update_available": {
-      const data = msg;
-      showUpdateCard(data.version, data.currentVersion, data.releaseNotes, data.downloadUrl);
-      break;
-    }
-
-    case "session_switched": {
-      const data = msg;
-
-      // Clear current state
-      state.entries = [];
-      state.currentTurn = null;
-      renderer.resetStreamingState();
-      renderer.clearMessages();
-      state.sessionStats = {};
-
-      // Apply the new state
-      if (data.state) {
-        state.model = data.state.model || null;
-        state.thinkingLevel = data.state.thinkingLevel || "off";
-        state.isStreaming = data.state.isStreaming || false;
-        state.isCompacting = data.state.isCompacting || false;
-        if (state.processStatus !== "crashed") state.processStatus = "running";
-      }
-
-      // Render historical messages
-      if (data.messages && data.messages.length > 0) {
-        renderHistoricalMessages(data.messages);
-      }
-
-      // Update session ID for the history panel
-      if (data.state?.sessionId) {
-        sessionHistory.setCurrentSessionId(data.state.sessionId as string);
-      }
-
-      // Hide history panel
-      sessionHistory.hide();
-
-      // Update all UI
-      updateAllUI();
-      scrollToBottom(messagesContainer, true);
-      break;
-    }
-  }
-
-  } catch (err: any) {
-    // Global error boundary — surface crashes visibly instead of silent failure
-    const errorId = `GSD-ERR-${Date.now().toString(36).toUpperCase()}`;
-    console.error(`[${errorId}] Message handler error for "${msg.type}":`, err);
-    addSystemEntry(
-      `Internal error processing "${msg.type}" (${errorId}). Check browser console for details. Please report this error code.`,
-      "error"
-    );
-  }
-});
-
-/**
- * Render historical messages from a switched session.
- * Converts the raw AgentMessage array into ChatEntry objects and renders them.
- *
- * Strategy: First pass collects tool results keyed by toolCallId. Second pass
- * builds entries, attaching tool results to their assistant turn's tool calls.
- */
-function renderHistoricalMessages(messages: import("../shared/types").AgentMessage[]): void {
-  // First pass: index tool results by toolCallId
-  const toolResults = new Map<string, { text: string; isError: boolean }>();
-  for (const msg of messages) {
-    if (msg.role === "toolResult") {
-      const toolCallId = (msg as Record<string, unknown>).toolCallId as string | undefined;
-      if (toolCallId) {
-        toolResults.set(toolCallId, {
-          text: extractMessageText(msg.content),
-          isError: !!(msg as Record<string, unknown>).isError,
-        });
-      }
-    }
-  }
-
-  // Second pass: render user and assistant messages
-  for (const msg of messages) {
-    if (msg.role === "user") {
-      const text = extractMessageText(msg.content);
-      if (!text) continue;
-      const entry: ChatEntry = {
-        id: nextId(),
-        type: "user",
-        text,
-        timestamp: msg.timestamp || Date.now(),
-      };
-      state.entries.push(entry);
-      renderer.renderNewEntry(entry);
-    } else if (msg.role === "assistant") {
-      const segments: import("./state").TurnSegment[] = [];
-      const turnToolCalls = new Map<string, ToolCallState>();
-
-      // Parse content blocks into segments
-      if (Array.isArray(msg.content)) {
-        for (const block of msg.content as Array<Record<string, unknown>>) {
-          if (block.type === "thinking" && block.thinking) {
-            segments.push({ type: "thinking", chunks: [block.thinking as string] });
-          } else if (block.type === "text" && block.text) {
-            const lastSeg = segments.length > 0 ? segments[segments.length - 1] : null;
-            if (lastSeg && lastSeg.type === "text") {
-              lastSeg.chunks.push(block.text as string);
-            } else {
-              segments.push({ type: "text", chunks: [block.text as string] });
-            }
-          } else if (block.type === "tool_use" && block.name) {
-            const toolId = (block.id as string) || nextId();
-            const result = toolResults.get(toolId);
-            const tc: ToolCallState = {
-              id: toolId,
-              name: block.name as string,
-              args: (block.input as Record<string, unknown>) || {},
-              resultText: result?.text || "",
-              isError: result?.isError || false,
-              isRunning: false,
-              startTime: msg.timestamp || Date.now(),
-              endTime: msg.timestamp || Date.now(),
-            };
-            turnToolCalls.set(toolId, tc);
-            segments.push({ type: "tool", toolCallId: toolId });
-          }
-        }
-      } else if (typeof msg.content === "string" && msg.content) {
-        segments.push({ type: "text", chunks: [msg.content] });
-      }
-
-      if (segments.length === 0) continue;
-
-      const turn = {
-        id: nextId(),
-        segments,
-        toolCalls: turnToolCalls,
-        isComplete: true,
-        timestamp: msg.timestamp || Date.now(),
-      };
-      const entry: ChatEntry = {
-        id: nextId(),
-        type: "assistant",
-        turn,
-        timestamp: msg.timestamp || Date.now(),
-      };
-      state.entries.push(entry);
-      renderer.renderNewEntry(entry);
-    }
-    // Skip toolResult (already indexed) and bashExecution
-  }
-
-  // Show messages area, hide welcome
-  if (state.entries.length > 0) {
-    welcomeScreen.style.display = "none";
-  }
-}
-
-/**
- * Extract text from a message content field (string or content array).
- */
-function extractMessageText(content: unknown): string {
-  if (typeof content === "string") return content;
-  if (Array.isArray(content)) {
-    return content
-      .filter((block: Record<string, unknown>) => block.type === "text")
-      .map((block: Record<string, unknown>) => block.text as string)
-      .filter(Boolean)
-      .join("\n");
-  }
-  return "";
-}
-
-/**
- * Show an inline update card in the chat with release notes and action buttons.
- */
-function showUpdateCard(
-  version: string,
-  currentVersion: string,
-  releaseNotes: string,
-  downloadUrl: string
-): void {
-  // Remove any existing update card
-  const existing = document.getElementById("gsd-update-card");
-  if (existing) existing.remove();
-
-
-
-  const card = document.createElement("div");
-  card.id = "gsd-update-card";
-  card.className = "gsd-update-card";
-  card.innerHTML = `
-    <div class="gsd-update-card-header">
-      <span class="gsd-update-icon">🚀</span>
-      <span class="gsd-update-title">Rokket GSD v${escapeHtml(version)} Available</span>
-      <span class="gsd-update-current">You have v${escapeHtml(currentVersion)}</span>
-    </div>
-    <div class="gsd-update-notes">
-      ${formatMarkdownNotes(releaseNotes)}
-    </div>
-    <div class="gsd-update-actions">
-      <button class="gsd-update-btn primary" data-action="install">Update Now</button>
-      <button class="gsd-update-btn dismiss" data-action="dismiss">Dismiss</button>
-    </div>
-  `;
-
-  // Wire up button handlers
-  card.querySelector('[data-action="install"]')?.addEventListener("click", () => {
-    vscode.postMessage({ type: "update_install", downloadUrl } as WebviewToExtensionMessage);
-    card.remove();
-  });
-
-  card.querySelector('[data-action="dismiss"]')?.addEventListener("click", () => {
-    vscode.postMessage({ type: "update_dismiss", version } as WebviewToExtensionMessage);
-    card.classList.add("dismissing");
-    setTimeout(() => card.remove(), 300);
-  });
-
-  // Insert at the top of the messages area, after any welcome screen
-  messagesContainer.insertBefore(card, messagesContainer.firstChild?.nextSibling || null);
-  scrollToBottom(messagesContainer);
-}
-
-/**
- * Show a "What's New" card on first launch after an update.
- */
-function showWhatsNew(version: string, notes: string): void {
-  const existing = document.getElementById("gsd-whats-new");
-  if (existing) existing.remove();
-
-  const card = document.createElement("div");
-  card.id = "gsd-whats-new";
-  card.className = "gsd-whats-new";
-  card.innerHTML = `
-    <div class="gsd-whats-new-header">
-      <span class="gsd-whats-new-icon">🚀</span>
-      <span class="gsd-whats-new-title">What's New in v${escapeHtml(version)}</span>
-      <button class="gsd-whats-new-close" title="Dismiss">✕</button>
-    </div>
-    <div class="gsd-whats-new-notes">
-      ${formatMarkdownNotes(notes)}
-    </div>
-  `;
-
-  card.querySelector(".gsd-whats-new-close")?.addEventListener("click", () => {
-    card.classList.add("dismissing");
-    setTimeout(() => card.remove(), 300);
-  });
-
-  messagesContainer.insertBefore(card, messagesContainer.firstChild?.nextSibling || null);
-  scrollToBottom(messagesContainer);
-}
-
-/**
- * Show a full changelog panel inline in the chat.
- */
-function showChangelog(entries: Array<{ version: string; notes: string; date: string }>): void {
-  const existing = document.getElementById("gsd-changelog");
-  if (existing) existing.remove();
-
-  const entriesHtml = entries.length > 0
-    ? entries.map((e, i) => `
-      <div class="gsd-changelog-entry${i === 0 ? " latest" : ""}">
-        <div class="gsd-changelog-entry-header">
-          <span class="gsd-changelog-version">v${escapeHtml(e.version)}</span>
-          ${i === 0 ? '<span class="gsd-changelog-latest-badge">latest</span>' : ""}
-          <span class="gsd-changelog-date">${formatShortDate(e.date)}</span>
-        </div>
-        <div class="gsd-changelog-entry-notes">${formatMarkdownNotes(e.notes)}</div>
-      </div>
-    `).join("")
-    : '<div class="gsd-changelog-empty">No changelog entries found.</div>';
-
-  const card = document.createElement("div");
-  card.id = "gsd-changelog";
-  card.className = "gsd-changelog";
-  card.innerHTML = `
-    <div class="gsd-changelog-header">
-      <span class="gsd-changelog-title">📋 Changelog</span>
-      <button class="gsd-changelog-close" title="Close">✕</button>
-    </div>
-    <div class="gsd-changelog-entries">
-      ${entriesHtml}
-    </div>
-  `;
-
-  card.querySelector(".gsd-changelog-close")?.addEventListener("click", () => {
-    card.classList.add("dismissing");
-    setTimeout(() => card.remove(), 300);
-  });
-
-  messagesContainer.appendChild(card);
-  scrollToBottom(messagesContainer, true);
-}
-
-function addSystemEntry(text: string, kind: "info" | "error" | "warning" = "info"): void {
-  const entry: ChatEntry = {
-    id: nextId(),
-    type: "system",
-    systemText: text,
-    systemKind: kind,
-    timestamp: Date.now(),
-  };
-  state.entries.push(entry);
-  renderer.renderNewEntry(entry);
-  scrollToBottom(messagesContainer);
-}
+// Message handler is in message-handler.ts
 
 // ============================================================
 // Initialize modules
 // ============================================================
+
+uiUpdates.init({
+  vscode,
+  modelBadge,
+  thinkingBadge,
+  headerSep1,
+  costBadge,
+  contextBadge,
+  contextBarContainer,
+  contextBar,
+  footerCwd,
+  footerStats,
+  footerRight,
+  sendBtn,
+  sendIcon,
+  promptInput,
+  inputHint,
+  overlayIndicators,
+});
+
+dashboard.init({
+  messagesContainer,
+  welcomeScreen,
+  welcomeProcess,
+  welcomeVersion,
+  welcomeModel,
+  welcomeHints,
+});
+
+fileHandling.init({
+  root,
+  imagePreview,
+  promptInput,
+  vscode,
+  onSendMessage: sendMessage,
+});
+
+keyboard.init({
+  vscode,
+  messagesContainer,
+  welcomeScreen,
+  promptInput,
+  sendBtn,
+  headerVersion,
+  newConvoBtn: newConvoBtn,
+  compactBtn,
+  exportBtn,
+  attachBtn,
+  thinkingBadge,
+  sendMessage,
+  updateAllUI,
+  autoResize,
+});
+
+messageHandler.init({
+  vscode,
+  messagesContainer,
+  welcomeScreen,
+  promptInput,
+  toolWatchdogTimers,
+  updateAllUI,
+  updateHeaderUI,
+  updateFooterUI,
+  updateInputUI,
+  updateOverlayIndicators,
+  updateWorkflowBadge,
+  autoResize,
+  announceToScreenReader,
+  startToolWatchdog,
+  clearToolWatchdog,
+});
 
 slashMenu.init({
   slashMenuEl: slashMenuEl,
@@ -2331,7 +611,7 @@ slashMenu.init({
   vscode,
   onAutoResize: autoResize,
   onShowModelPicker: modelPicker.show,
-  onNewConversation: handleNewConversation,
+  onNewConversation: keyboard.handleNewConversation,
   onSendMessage: sendMessage,
   onShowHistory: () => {
     vscode.postMessage({ type: "get_session_list" });
@@ -2349,19 +629,19 @@ slashMenu.init({
         const text = textChunks.join("\n");
         if (text) {
           vscode.postMessage({ type: "copy_text", text });
-          addSystemEntry("Last assistant message copied to clipboard.", "info");
+          messageHandler.addSystemEntry("Last assistant message copied to clipboard.", "info");
           return;
         }
       }
     }
-    addSystemEntry("No assistant message to copy.", "warning");
+    messageHandler.addSystemEntry("No assistant message to copy.", "warning");
   },
   onToggleAutoCompact: () => {
     const current = state.sessionStats.autoCompactionEnabled;
     const newValue = !current;
     vscode.postMessage({ type: "set_auto_compaction", enabled: newValue });
     state.sessionStats.autoCompactionEnabled = newValue;
-    addSystemEntry(`Auto-compaction ${newValue ? "enabled" : "disabled"}.`, "info");
+    messageHandler.addSystemEntry(`Auto-compaction ${newValue ? "enabled" : "disabled"}.`, "info");
   },
 });
 
@@ -2391,7 +671,7 @@ sessionHistory.init({
   onSessionSwitched: () => {
     updateAllUI();
   },
-  onNewConversation: handleNewConversation,
+  onNewConversation: keyboard.handleNewConversation,
 });
 
 uiDialogs.init({
